% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ancovaN}
\alias{ancovaN}
\title{Ancova n Function}
\usage{
ancovaN(pow = 0.9, del = 0.4, r2 = 0.5, alp = 0.05)
}
\arguments{
\item{simple}{parats.}
}
\description{
This function allows you to find n.
}
\examples{
ancovaN()
}
\keyword{ancova}

